+++ src/main/java/kitchenpos/products/domain/ProfanityValidator.java
@@ -1,0 +1,5 @@
+package kitchenpos.products.domain;
+
+public interface ProfanityValidator {
+    boolean isHas(String word);
+}
